requirements.txt
django
djangorestframework
markdown
django-filter
pymysql
configparser
django-tables2
django-crispy-forms
django-bootstrap3
django-filters
django-crowdsourcing
django-countries
django-formtools
